Refactor greedy and enum

+++ comp60342/knapsack/readsack.py
@@ -8,8 +8,8 @@ class Sack(object):
     """A sack containing items, has certain capacity"""
 
     def __init__(self, num=0, capacity=0, items=None):
-        self.__num = num
-        self.__capacity = capacity
+        self.__num = int(num)
+        self.__capacity = int(capacity)
         if items is None:
             self.__items = []
         else:
@@ -19,13 +19,13 @@ class Sack(object):
         return self.__capacity
 
     def set_capacity(self, capacity):
-        self.__capacity = capacity
+        self.__capacity = int(capacity)
 
     def get_num(self):
         return self.__num
 
     def set_num(self, num):
-        self.__num = num
+        self.__num = int(num)
 
     def pour(self):
         return self.__items
@@ -39,8 +39,8 @@ class Item(object):
 
     def __init__(self, *args):
         self.name = args[0]
-        self.value = float(args[1])
-        self.weight = float(args[2])
+        self.value = int(args[1])
+        self.weight = int(args[2])
 
 
 def readsack(problem):
@@ -49,12 +49,12 @@ def readsack(problem):
     sack = Sack()
 
     with open(problem, 'r') as reader:
-        sack.set_num(int(reader.readline().rstrip()))
+        sack.set_num(reader.readline().rstrip())
         for line in reader:
             try:
                 attributes = line.split()
                 sack.push(Item(*attributes))
             except:
-                sack.set_capacity(int(attributes[0]))
+                sack.set_capacity(attributes[0])
 
     return sack
